backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.2
openai==1.45.0
google-api-python-client==2.146.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
apscheduler==3.10.4
pandas==2.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.2
openpyxl==3.1.5
cloudinary==1.41.0
